+++ spec/rubocop/cop/metrics/line_length_spec.rb
@@ -410,4 +410,504 @@ RSpec.describe RuboCop::Cop::Metrics::LineLength, :config do
       end
     end
   end
+
+  context 'autocorrection' do
+    let(:cop_config) do
+      { 'Max' => 40, 'IgnoredPatterns' => nil, 'AutoCorrect' => true }
+    end
+
+    context 'hash' do
+      context 'when under limit' do
+        it 'does not add any offenses' do
+          expect_no_offenses(<<-RUBY.strip_indent)
+            {foo: 1, bar: "2"}
+          RUBY
+        end
+      end
+
+      context 'when over limit because of a comment' do
+        it 'adds offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            { # supersupersupersupersupersupersupersupersupersupersupersuperlongcomment
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [75/40]
+              baz: "10000",
+              bar: "10000"}
+          RUBY
+        end
+
+        it 'does not autocorrect the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            { # supersupersupersupersupersupersupersupersupersupersupersuperlongcomment
+              baz: "10000",
+              bar: "10000"}
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            { # supersupersupersupersupersupersupersupersupersupersupersuperlongcomment
+              baz: "10000",
+              bar: "10000"}
+          RUBY
+        end
+      end
+
+      context 'when over limit and already on multiple lines long key' do
+        it 'adds offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            {supersupersupersupersupersupersupersupersupersupersupersuperfirstarg: 10,
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [74/40]
+              baz: "10000",
+              bar: "10000"}
+          RUBY
+        end
+
+        it 'does not autocorrect the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            {supersupersupersupersupersupersupersupersupersupersupersuperfirstarg: 10,
+              baz: "10000",
+              bar: "10000"}
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            {supersupersupersupersupersupersupersupersupersupersupersuperfirstarg: 10,
+              baz: "10000",
+              bar: "10000"}
+          RUBY
+        end
+      end
+
+      context 'when over limit and keys already on multiple lines' do
+        it 'adds offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            {
+              baz0: "10000",
+              baz1: "10000",
+              baz2: "10000", baz2: "10000", baz3: "10000", baz4: "10000",
+                                                    ^^^^^^^^^^^^^^^^^^^^^ Line is too long. [61/40]
+              bar: "10000"}
+          RUBY
+        end
+
+        it 'does not autocorrect the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            {
+              baz0: "10000",
+              baz1: "10000",
+              baz2: "10000", baz2: "10000", baz3: "10000", baz4: "10000",
+              bar: "10000"}
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            {
+              baz0: "10000",
+              baz1: "10000",
+              baz2: "10000", baz2: "10000", baz3: "10000", baz4: "10000",
+              bar: "10000"}
+          RUBY
+        end
+      end
+
+      context 'when over limit' do
+        it 'adds an offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            {abc: "100000", def: "100000", ghi: "100000", jkl: "100000", mno: "100000"}
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [75/40]
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            {abc: "100000", def: "100000", ghi: "100000", jkl: "100000", mno: "100000"}
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            {abc: "100000", def: "100000",\s
+            ghi: "100000", jkl: "100000", mno: "100000"}
+          RUBY
+        end
+      end
+
+      context 'when over limit rocket' do
+        it 'adds an offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            {"abc" => "100000", "def" => "100000", "casd" => "100000", "asdf" => "100000"}
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [78/40]
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            {"abc" => "100000", "def" => "100000", "casd" => "100000", "asdf" => "100000"}
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            {"abc" => "100000", "def" => "100000",\s
+            "casd" => "100000", "asdf" => "100000"}
+          RUBY
+        end
+      end
+
+      context 'when over limit rocket symbol' do
+        it 'adds an offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            {:abc => "100000", :asd => "100000", :asd => "100000", :fds => "100000"}
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [72/40]
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            {:abc => "100000", :asd => "100000", :asd => "100000", :fds => "100000"}
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            {:abc => "100000", :asd => "100000",\s
+            :asd => "100000", :fds => "100000"}
+          RUBY
+        end
+      end
+
+      context 'when nested hashes on same line' do
+        it 'adds an offense only to outer' do
+          expect_offense(<<-RUBY.strip_indent)
+            {abc: "100000", def: "100000", ghi: {abc: "100000"}, jkl: "100000", mno: "100000"}
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [82/40]
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            {abc: "100000", def: "100000", ghi: {abc: "100000"}, jkl: "100000", mno: "100000"}
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            {abc: "100000", def: "100000",\s
+            ghi: {abc: "100000"}, jkl: "100000", mno: "100000"}
+          RUBY
+        end
+      end
+
+      context 'when hash in method call' do
+        it 'adds an offense only to outer' do
+          expect_offense(<<-RUBY.strip_indent)
+            get(
+              :index,
+              params: {driver_id: driver.id, from_date: "2017-08-18T15:09:04.000Z", to_date: "2017-09-19T15:09:04.000Z"},
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [109/40]
+              xhr: true)
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            get(
+              :index,
+              params: {driver_id: driver.id, from_date: "2017-08-18T15:09:04.000Z", to_date: "2017-09-19T15:09:04.000Z"},
+              xhr: true)
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            get(
+              :index,
+              params: {driver_id: driver.id,\s
+            from_date: "2017-08-18T15:09:04.000Z", to_date: "2017-09-19T15:09:04.000Z"},
+              xhr: true)
+          RUBY
+        end
+      end
+    end
+
+    context 'method call' do
+      context 'when under limit' do
+        it 'does not add any offenses' do
+          expect_no_offenses(<<-RUBY.strip_indent)
+            foo(foo: 1, bar: "2")
+          RUBY
+        end
+      end
+
+      context 'when two together' do
+        it 'does not add any offenses' do
+          expect_no_offenses(<<-RUBY.strip_indent)
+            def baz(bar)
+              foo(shipment, actionable_delivery) &&
+                bar(shipment, actionable_delivery)
+            end
+          RUBY
+        end
+      end
+
+      context 'when over limit' do
+        it 'adds an offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            foo(abc: "100000", def: "100000", ghi: "100000", jkl: "100000", mno: "100000")
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [78/40]
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            foo(abc: "100000", def: "100000", ghi: "100000", jkl: "100000", mno: "100000")
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            foo(abc: "100000", def: "100000",\s
+            ghi: "100000", jkl: "100000", mno: "100000")
+          RUBY
+        end
+      end
+
+      context 'when call with hash on same line' do
+        it 'adds an offense only to outer' do
+          expect_offense(<<-RUBY.strip_indent)
+              foo(abc: "100000", def: "100000", ghi: {abc: "100000"}, jkl: "100000", mno: "100000")
+                                                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [85/40]
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            foo(abc: "100000", def: "100000", ghi: {abc: "100000"}, jkl: "100000", mno: "100000")
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            foo(abc: "100000", def: "100000",\s
+            ghi: {abc: "100000"}, jkl: "100000", mno: "100000")
+          RUBY
+        end
+      end
+
+      context 'when two method calls' do
+        it 'adds an offense only to outer' do
+          expect_offense(<<-RUBY.strip_indent)
+            get(1000000, 30000, foo(44440000, 30000, 39999, 19929120312093))
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [64/40]
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            get(1000000, 30000, foo(44440000, 30000, 39999, 19929120312093))
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            get(1000000, 30000,\s
+            foo(44440000, 30000, 39999, 19929120312093))
+          RUBY
+        end
+      end
+
+      context 'when nested method calls allows outer to get broken up first' do
+        it 'adds offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            get(1000000,
+            foo(44440000, 30000, 39999, 1992), foo(44440000, 30000, 39999, 12093))
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [70/40]
+          RUBY
+        end
+
+        it 'does not autocorrect the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            get(1000000,
+            foo(44440000, 30000, 39999, 1992), foo(44440000, 30000, 39999, 12093))
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            get(1000000,
+            foo(44440000, 30000, 39999, 1992), foo(44440000, 30000, 39999, 12093))
+          RUBY
+        end
+      end
+    end
+
+    context 'array' do
+      context 'when under limit' do
+        it 'does not add any offenses' do
+          expect_no_offenses(<<-RUBY.strip_indent)
+            [1, "2"]
+          RUBY
+        end
+      end
+
+      context 'when already on two lines' do
+        it 'does not add any offenses' do
+          expect_no_offenses(<<-RUBY.strip_indent)
+            [1, "2",
+             "3"]
+          RUBY
+        end
+      end
+
+      context 'when over limit' do
+        it 'adds an offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            ["1111", "100000", "100000", "100000", "100000", "100000"]
+                                                    ^^^^^^^^^^^^^^^^^^ Line is too long. [58/40]
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            ["1111", "100000", "100000", "100000", "100000", "100000"]
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            ["1111", "100000", "100000", "100000",\s
+            "100000", "100000"]
+          RUBY
+        end
+      end
+
+      context 'when has inside array' do
+        it 'adds an offense only to outer' do
+          expect_offense(<<-RUBY.strip_indent)
+            ["1111", "100000", "100000", "100000", {abc: "100000", b: "2"}, "100000", "100000"]
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [83/40]
+          RUBY
+        end
+
+        it 'autocorrects the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            ["1111", "100000", "100000", "100000", {abc: "100000", b: "2"}, "100000", "100000"]
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            ["1111", "100000", "100000", "100000",\s
+            {abc: "100000", b: "2"}, "100000", "100000"]
+          RUBY
+        end
+      end
+
+      context 'when two arrays on two lines allows outer to get broken first' do
+        it 'adds an offense only to inner' do
+          expect_offense(<<-RUBY.strip_indent)
+            [1000000, 3912312312999,
+              [44440000, 3912312312999, 3912312312999, 1992912031231232131312093],
+                                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [70/40]
+            100, 100]
+          RUBY
+        end
+
+        it 'does not autocorrect the offense' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            [1000000, 3912312312999,
+              [44440000, 3912312312999, 3912312312999, 1992912031231232131312093],
+            100, 100]
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            [1000000, 3912312312999,
+              [44440000, 3912312312999, 3912312312999, 1992912031231232131312093],
+            100, 100]
+          RUBY
+        end
+      end
+    end
+
+    context 'no breakable collections' do
+      it 'adds an offense' do
+        expect_offense(<<-RUBY.strip_indent)
+          10000003912312312999
+            # 444400003912312312999391231231299919929120312312321313120933333333
+                                                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Line is too long. [70/40]
+          456
+        RUBY
+      end
+
+      it 'does not autocorrect the offense' do
+        new_source = autocorrect_source(<<-RUBY.strip_indent)
+          10000003912312312999
+            # 444400003912312312999391231231299919929120312312321313120933333333
+          456
+        RUBY
+
+        expect(new_source).to eq(<<-RUBY.strip_indent)
+          10000003912312312999
+            # 444400003912312312999391231231299919929120312312321313120933333333
+          456
+        RUBY
+      end
+    end
+
+    context 'semicolon' do
+      context 'when under limit' do
+        it 'does not add any offenses' do
+          expect_no_offenses(<<-RUBY.strip_indent)
+            {foo: 1, bar: "2"}; a = 4 + 5
+          RUBY
+        end
+      end
+
+      context 'when over limit' do
+        it 'adds offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            {foo: 1, bar: "2"}; a = 400000000000 + 500000000000000
+                                                    ^^^^^^^^^^^^^^ Line is too long. [54/40]
+          RUBY
+        end
+
+        it 'breaks the semicolon before breaking the hash' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            {foo: 1, bar: "2"}; a = 400000000000 + 500000000000000
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            {foo: 1, bar: "2"};
+             a = 400000000000 + 500000000000000
+          RUBY
+        end
+      end
+    end
+
+    context 'HEREDOC' do
+      let(:cop_config) do
+        { 'Max' => 40, 'AllowURI' => false, 'AllowHeredoc' => false }
+      end
+
+      context 'when over limit with semi-colon' do
+        it 'adds offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            foo = <<-SQL
+              SELECT a b c d a b FROM c d a b c d ; COUNT(*) a b
+                                       ^^^^^^^^^^^^ Line is too long. [52/40]
+            SQL
+          RUBY
+        end
+
+        it 'does not autocorrect' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            foo = <<-SQL
+              SELECT a b c d a b FROM c d a b c d ; COUNT(*) a b c
+            SQL
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            foo = <<-SQL
+              SELECT a b c d a b FROM c d a b c d ; COUNT(*) a b c
+            SQL
+          RUBY
+        end
+      end
+    end
+
+    context 'comments' do
+      context 'when over limit with semi-colon' do
+        it 'adds offense' do
+          expect_offense(<<-RUBY.strip_indent)
+            # a b c d a b c d a b c d ; a b c d a b c d a b c d a
+                                                    ^^^^^^^^^^^^^ Line is too long. [53/40]
+          RUBY
+        end
+
+        it 'does not autocorrect' do
+          new_source = autocorrect_source(<<-RUBY.strip_indent)
+            # a b c d a b c d a b c d ; a b c d a b c d a b c d a
+          RUBY
+
+          expect(new_source).to eq(<<-RUBY.strip_indent)
+            # a b c d a b c d a b c d ; a b c d a b c d a b c d a
+          RUBY
+        end
+      end
+    end
+  end
 end